pyproject.toml
[tool.poetry]
name = "cacheback_test"
version = "0.1.0"
description = ""
authors = ["Carlos Gottberg <13184659+el-chogo@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
celery = "^5.2.7"
django-cacheback = {extras = ["celery"], version = "^3.0.0"}
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
